refactor(tools): localize schemas and document boundaries
